requirements.txt
# Azure SDK for interacting with Azure Blob Storage
azure-storage-blob
azure-identity
azure-keyvault-secrets

# Libraries for data manipulation
pandas
pyarrow # For Parquet support

# Libraries for connecting to Azure SQL Database
pyodbc
sqlmodel

duckdb

python-dotenv